chore: segregate inteface
