--- sd/h2/src/main/java/ru/bauman/tigerbank/account/service/balance/BalanceRecalculationServiceInterface.java
@@ -1,8 +1,0 @@
-package ru.bauman.tigerbank.account.service.balance;
-
-import java.math.BigDecimal;
-
-public interface BalanceRecalculationServiceInterface {
-	void autoRecalc(Long accountId);
-	void manualRecalc(Long accountId, BigDecimal correctBalance);
-}
