+++ sd/h2/src/main/java/ru/bauman/tigerbank/account/service/balance/impl/BalanceRecalculationServiceImpl.java
@@ -1,0 +1,57 @@
+package ru.bauman.tigerbank.account.service.balance.impl;
+
+import ru.bauman.tigerbank.account.entity.BankAccount;
+import ru.bauman.tigerbank.account.service.balance.BalanceRecalculationService;
+import ru.bauman.tigerbank.operation.entity.Operation;
+import ru.bauman.tigerbank.operation.entity.OperationType;
+import ru.bauman.tigerbank.operation.entity.OperationTypeEnum;
+import ru.bauman.tigerbank.account.service.entity.BankAccountEntityService;
+import ru.bauman.tigerbank.operation.service.entity.OperationEntityService;
+import ru.bauman.tigerbank.operation.service.entity.OperationTypeEntityServiceInterface;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import java.math.BigDecimal;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class BalanceRecalculationServiceImpl implements BalanceRecalculationService {
+
+	private final BankAccountEntityService accountEntityService;
+	private final OperationEntityService operationEntityService;
+	private final OperationTypeEntityServiceInterface operationTypeEntityService;
+
+	@Override
+	@Transactional
+	public void autoRecalc(Long accountId) {
+		BankAccount account = accountEntityService.getById(accountId);
+		OperationType incomeType = operationTypeEntityService.getByName(OperationTypeEnum.INCOME);
+		OperationType expenseType = operationTypeEntityService.getByName(OperationTypeEnum.EXPENSE);
+		List<Operation> operations = operationEntityService.findAllByAccountId(accountId);
+		BigDecimal balance = operations.stream()
+				.map(op -> {
+					if (op.getType().getId().equals(incomeType.getId())) {
+						return op.getAmount();
+					} else if (op.getType().getId().equals(expenseType.getId())) {
+						return op.getAmount().negate();
+					} else {
+						return BigDecimal.ZERO;
+					}
+				})
+				.reduce(BigDecimal.ZERO, BigDecimal::add);
+		account.setBalance(balance);
+		accountEntityService.save(account);
+	}
+
+	@Override
+	@Transactional
+	public void manualRecalc(Long accountId, BigDecimal correctBalance) {
+		autoRecalc(accountId);
+		BankAccount account = accountEntityService.getById(accountId);
+		if (account.getBalance().compareTo(correctBalance) != 0) { // validate balance
+			account.setBalance(correctBalance);
+			accountEntityService.save(account);
+		}
+	}
+}
