+++ sd/h2/src/main/java/ru/bauman/tigerbank/account/service/entity/BankAccountEntityService.java
@@ -3,7 +3,7 @@ package ru.bauman.tigerbank.account.service.entity;
 import ru.bauman.tigerbank.account.entity.BankAccount;
 import java.util.List;
 
-public interface BankAccountEntityServiceInterface {
+public interface BankAccountEntityService {
 	BankAccount save(BankAccount account);
 	void deleteById(Long id);
 	BankAccount getById(Long id);
