+++ sd/h2/src/main/java/ru/bauman/tigerbank/account/service/impl/BankAccountService.java
@@ -4,7 +4,7 @@ import ru.bauman.tigerbank.account.service.BankAccountServiceInterface;
 import ru.bauman.tigerbank.account.dto.BankAccountDto;
 import ru.bauman.tigerbank.account.mapper.BankAccountMapper;
 import ru.bauman.tigerbank.account.entity.BankAccount;
-import ru.bauman.tigerbank.account.service.entity.BankAccountEntityServiceInterface;
+import ru.bauman.tigerbank.account.service.entity.BankAccountEntityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,7 +13,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class BankAccountService implements BankAccountServiceInterface {
-	private final BankAccountEntityServiceInterface entityService;
+	private final BankAccountEntityService entityService;
 	private final BankAccountMapper mapper;
 
 	@Override
