+++ sd/h2/src/main/java/ru/bauman/tigerbank/operation/service/entity/OperationEntityService.java
@@ -4,11 +4,12 @@ import ru.bauman.tigerbank.operation.entity.Operation;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface OperationEntityServiceInterface {
+public interface OperationEntityService {
 	Operation save(Operation operation);
 	void deleteById(Long id);
 	Operation getById(Long id);
 	List<Operation> getAll();
 	List<Operation> findByAccountAndPeriod(Long accountId, LocalDateTime from, LocalDateTime to);
 	List<Operation> findByCategory(Long categoryId);
+	List<Operation> findAllByAccountId(Long accountId);
 }
